requirements.txt
neat-python==0.92
numpy==1.14.3
gym
graphviz
matplotlib
torch==0.4.1
